chore: slight refactor
