+++ src/components/pages/blog/TagsFiltersWidget.tsx
@@ -69,7 +69,7 @@ const TagsFiltersWidget: FunctionComponent<TagsFiltersWidgetProps> = ({
   useEffect(() => {
     function unsafeCtxHandler() {
       try {
-        const unpackedAndCleanedFilters = getUnpackedAndSanitizedFilters(searchParams);
+        const unpackedAndCleanedFilters = getUnpackedAndSanitizedFilters(searchParams.get(FILTERS_KEY));
         setSelectedTagsIds(unpackedAndCleanedFilters);
 
         const sanitizedFilters = packIds(unpackedAndCleanedFilters);
