+++ src/components/pages/blog/helpers/functions.ts
@@ -3,15 +3,12 @@ import type { MaybeNull } from '@rtm/shared-types/CustomUtilityTypes';
 import { blogTagOptions } from '##/lib/builders/unifiedImport';
 import { unpackIds } from '@rtm/shared-lib/misc';
 
-import { FILTERS_KEY } from './constants';
-
 export const sortUnpackedIds = (unpacked: number[]) => unpacked.sort((a, b) => a - b);
 
 /**
  * @throws {RangeError}
  */
-export function getUnpackedAndSanitizedFilters(searchParams: URLSearchParams, filtersKey: string = FILTERS_KEY) {
-  const packedIds: MaybeNull<string> = searchParams.get(filtersKey);
+export function getUnpackedAndSanitizedFilters(packedIds: MaybeNull<string>) {
   if (packedIds === null) return [];
 
   const unpackedAndSanitizedFilters = sortUnpackedIds(
